fix: resolve all clippy warnings, remove Cesium Ion branding, update test counts
- Fix dead_code warning for QuantizedMeshHeader in tiletopia-terrain
- Fix unnecessary u32 cast in tiletopia-terrain
- Fix unused variable in streaming.rs
- Fix io_other_error in annotations.rs and temporal.rs
- Add Default impl for RealtimeState
- Remove Cesium Ion token and hide credit container in viewer
- Update test count badges from 29 to 53 in docs and README

+++ crates/tiletopia-terrain/src/lib.rs
@@ -90,6 +90,7 @@ pub struct QuantizedMeshTile {
 
 /// Quantized mesh header (Cesium terrain format).
 #[repr(C, packed)]
+#[allow(dead_code)]
 struct QuantizedMeshHeader {
     center_x: f64,
     center_y: f64,
@@ -141,7 +142,7 @@ pub fn generate_quantized_mesh(heightmap: &Heightmap) -> Vec<u8> {
     }
 
     // Build output buffer
-    let vertex_count = (w * h) as u32;
+    let vertex_count = w * h;
     let mut buf = Vec::new();
 
     // Header (simplified — real implementation needs ECEF coords)
